refactor: consolidate domain controllers
